Fixed mixmat read bug.

+++ Processor/gaussian.cpp
@@ -15,13 +15,13 @@ namespace handinput {
   }
 
   // p(x) = 1/sqrt((2pi)^k * det(cov))*exp(-0.5(x - mean)'cov^-1(x - mean))
-  float Gaussian::Prob(const Eigen::Ref<const Eigen::VectorXf> x) const {
+  double Gaussian::Prob(const Eigen::Ref<const Eigen::VectorXf> x) const {
     using Eigen::VectorXf;
     if (x.size() != mean_.size())
       throw std::invalid_argument("The size of the input is incorrect.");
     VectorXf v = x - mean_;
     float mahal = (inv_cov_ * v).dot(v);
-    return (float) (b_ * exp(-0.5 * mahal));
+    return b_ * exp(-0.5 * mahal);
   }
 
   void Gaussian::CheckRI() {
